requirements.txt
Chameleon==2.24
Mako==1.0.3
MarkupSafe==0.23
PasteDeploy==1.5.2
Pygments==2.0.2
SQLAlchemy==1.0.9
WebOb==1.5.1
argparse==1.2.1
decorator==4.0.4
ipdb==0.8.1
ipython==4.0.0
ipython-genutils==0.1.0
nose==1.3.7
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
pyramid==1.5.7
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.4.2
pyramid-mako==1.0.2
pyramid-tm==0.12
python-dateutil==2.4.2
python-editor==0.4
pytz==2015.7
repoze.lru==0.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.0.0
transaction==1.4.4
translationstring==1.3
venusian==1.0
waitress==0.8.10
wsgiref==0.1.2
zope.deprecation==4.1.2
zope.interface==4.1.3
zope.sqlalchemy==0.7.6
